--- a/cmake/external/composable_kernel.cmake
+++ b/cmake/external/composable_kernel.cmake
@@ -10,6 +10,7 @@ FetchContent_Declare(composable_kernel
 FetchContent_GetProperties(composable_kernel)
 if(NOT composable_kernel_POPULATED)
   FetchContent_Populate(composable_kernel)
+  set(GPU_TARGETS ${CMAKE_HIP_ARCHITECTURES})
   set(BUILD_DEV OFF CACHE BOOL "Disable -Weverything, otherwise, error: 'constexpr' specifier is incompatible with C++98 [-Werror,-Wc++98-compat]" FORCE)
   # Exclude i8 device gemm instances due to excessive long compilation time and not being used
   set(DTYPES fp32 fp16 bf16 fp8)