AMD claim improved gemm performance. update CK to see perf.

llama2 7b decoding shape:
```
old:
 75.44 us 21.35 tflops 674.14 GB/s float16_float8_e4m3fn_float16 NN m=16   n=12288 k=4096 Tunable *
 34.61 us 15.51 tflops 492.28 GB/s float16_float8_e4m3fn_float16 NN m=16   n=4096 k=4096 Tunable *
139.33 us 20.71 tflops 653.21 GB/s float16_float8_e4m3fn_float16 NN m=16   n=22016 k=4096 Tunable *
112.90 us 12.78 tflops 403.66 GB/s float16_float8_e4m3fn_float16 NN m=16   n=4096 k=11008 Tunable *

new:
 76.58 us 21.03 tflops 664.08 GB/s float16_float8_e4m3fn_float16 NN m=16   n=12288 k=4096 Tunable *
 33.90 us 15.84 tflops 502.65 GB/s float16_float8_e4m3fn_float16 NN m=16   n=4096 k=4096 Tunable *
141.70 us 20.37 tflops 642.31 GB/s float16_float8_e4m3fn_float16 NN m=16   n=22016 k=4096 Tunable *
116.19 us 12.42 tflops 392.21 GB/s float16_float8_e4m3fn_float16 NN m=16   n=4096 k=11008 Tunable *
```

llama2 7b prefill shape:
```
old:
928.29 us 111.04 tflops 90.37 GB/s float16_float8_e4m3fn_float16 NN m=1024 n=12288 k=4096 Tunable *
357.80 us 96.03 tflops 93.78 GB/s float16_float8_e4m3fn_float16 NN m=1024 n=4096 k=4096 Tunable *
1700.44 us 108.61 tflops 84.48 GB/s float16_float8_e4m3fn_float16 NN m=1024 n=22016 k=4096 Tunable *
922.47 us 100.10 tflops 82.41 GB/s float16_float8_e4m3fn_float16 NN m=1024 n=4096 k=11008 Tunable *

new:
908.77 us 113.43 tflops 92.31 GB/s float16_float8_e4m3fn_float16 NN m=1024 n=12288 k=4096 Tunable *
365.73 us 93.95 tflops 91.75 GB/s float16_float8_e4m3fn_float16 NN m=1024 n=4096 k=4096 Tunable *
1665.39 us 110.90 tflops 86.26 GB/s float16_float8_e4m3fn_float16 NN m=1024 n=22016 k=4096 Tunable *
903.62 us 102.19 tflops 84.13 GB/s float16_float8_e4m3fn_float16 NN m=1024 n=4096 k=11008 Tunable *
```

No obvious improvement. Will hold off.